requirements.txt
argcomplete==1.10.3
beautifulsoup4==4.8.2
boto3==1.25.3
botocore==1.28.3
chardet==3.0.4
click==8.1.3
colorama==0.4.6
compressed-rtf==1.0.6
dnspython==2.2.1
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
Flask==2.2.2
Flask-Cors==3.0.10
IMAPClient==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
lxml==4.9.1
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.4
olefile==0.46
pandas==1.5.1
pdfminer.six==20191110
Pillow==9.2.0
pycryptodome==3.15.0
pymongo==4.3.3
PyMuPDF==1.20.2
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
regex==2022.9.13
s3transfer==0.6.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
textract==1.6.5
textract-trp==0.1.3
tqdm==4.64.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
xlrd==1.2.0
XlsxWriter==3.0.3
